+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Bug fixes and minor improvements
 - Update `px_micro()` documentation.
+- Bugfix: `px_add_totals()` now uses default value 'Total' if 
+`x$variables2$elimination` is `NA`. (#284)
 
 # pxmake 0.11.0